backend/requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
python-nmap>=0.7.1
python-dotenv>=0.19.0
requests==2.31.0
gunicorn==21.2.0
pytest==7.4.0
werkzeug>=2.0.1
psutil>=5.9.0  # 프로세스 관리 및 시스템 정보 수집에 사용 